Create README.md
i am a doctor by profession who is looking to learn R language as a side hobby. i have to look at different stats of common diseases for research purposes and data analysis , bio stats and the data science becomes mandatory to compile the vast data we gather. learning this language will give me an edge and an advantage.